+++ vsc-base.org/src/allAnnotations/annotations/TsInsertImportAnnotatedCode.tsx
@@ -58,17 +58,17 @@ export const tsInsertImport: (
    if (matchImport) \{
       return source
    }
-   const [matchImportPath] = vsc.tsFindNodePositionFromContent(source, node =>
+   const [matchImportPath, matchImportPathPos] = vsc.tsFindNodePositionFromContent(source, node =>
       vsc.tsMatchImport(node, \{
          path: importPath
       })
    )
-   if (matchImportPath) \{
+   if (matchImportPath && matchImportPathPos) \{
       let importContent = matchImportPath.getText()
       importContent = isDefault
          ? importContent.replace('import ', \`import \$\{importName}, \`)
          : importContent.replace('import \{', \`import \{ \$\{importName},\`)
-      source = source.substring(0, matchImportPath.pos) + importContent + source.substring(matchImportPath.end);
+      source = source.substring(0, matchImportPathPos.start) + importContent + source.substring(matchImportPathPos.end);
       return source
    }