+++ vsc-base/out/vsc-base-typescript-top-level-transform.js
@@ -28,15 +28,15 @@ exports.tsInsertImport = (source, importName, importPath, options) => {
     if (matchImport) {
         return source;
     }
-    const [matchImportPath] = vsc.tsFindNodePositionFromContent(source, node => vsc.tsMatchImport(node, {
+    const [matchImportPath, matchImportPathPos] = vsc.tsFindNodePositionFromContent(source, node => vsc.tsMatchImport(node, {
         path: importPath
     }));
-    if (matchImportPath) {
+    if (matchImportPath && matchImportPathPos) {
         let importContent = matchImportPath.getText();
         importContent = isDefault
             ? importContent.replace('import ', `import ${importName}, `)
             : importContent.replace('import {', `import { ${importName},`);
-        source = source.substring(0, matchImportPath.pos) + importContent + source.substring(matchImportPath.end);
+        source = source.substring(0, matchImportPathPos.start) + importContent + source.substring(matchImportPathPos.end);
         return source;
     }
     const allImports = vsc.tsFindAllNodePositionsFromContent(source, node => vsc.tsMatchImport(node)).map(([imp, pos]) => imp);
@@ -186,7 +186,7 @@ exports.tsInsertInterfaceMember = (source, interfaceName, memberName, type, opti
         return source;
     }
     // check that the property don't exist
-    const hasMember = _interface.members.find(m => m.name && m.name.getText() === memberName);
+    const hasMember = _interface.members.find(m => !!m.name && m.name.getText() === memberName);
     if (hasMember) {
         return source;
     }
@@ -274,7 +274,7 @@ exports.tsInsertVariableObjectProperty = (source, variableName, key, value, opti
         return source;
     }
     // check that the property don't exist
-    const hasValue = obj.properties.find(p => p.name && p.name.getText() === key);
+    const hasValue = obj.properties.find(p => !!p.name && p.name.getText() === key);
     if (hasValue) {
         return source;
     }
@@ -323,10 +323,17 @@ exports.tsInsertVariableObjectProperty = (source, variableName, key, value, opti
     else {
         contentBeforeNewProp = contentBeforeNewProp + newIntentionString;
     }
+    let newContent;
+    if (value) {
+        newContent = `${contentBeforeNewProp}${key}: ${value}${contentAfterNewProp}`;
+    }
+    else {
+        newContent = `${contentBeforeNewProp}${key}${contentAfterNewProp}`;
+    }
     // Add property
     source =
         source.substring(0, insertPoint) +
-            `${contentBeforeNewProp}${key}: ${value}${contentAfterNewProp}` +
+            newContent +
             source.substring(insertPoint);
     // Add comma after prev property
     if (hasProps && !leadingComma) {