+++ vsc-script/out/vsc-base-development/vsc-base-typescript-match.js
@@ -941,6 +941,27 @@ exports.tsIsVariable = (node, options) => {
 };
 /** vsc-base method
  * @description
+ * Test is a node is a variable declaration (node: ts.VariableDeclarationList) \
+ * Uses [tsMatchVariableList](http://vsc-base.org/#tsMatchVariableList)
+ * @see [tsIsVariableList](http://vsc-base.org/#tsIsVariableList)
+ * @vscType ts
+ * @example
+ * const isVariableNode = vsc.tsIsVariableList(node, options)
+ * @example
+ * const isVariableNode = vsc.tsIsVariableList(
+ *   node, {
+ *     hasVariable: variable =>
+ *       variable.name.getText() === variableName
+ *       && ts.isObjectLiteralExpression(variable.initializer)
+ *   }
+ * )
+ * @returns boolean
+ */
+exports.tsIsVariableList = (node, options) => {
+    return !!vsc.tsMatchVariableList(node, options);
+};
+/** vsc-base method
+ * @description
  * Test is a node is a variable declaration (node: ts.VariableDeclaration) \
  * Uses [tsMatchIdentifier](http://vsc-base.org/#tsMatchIdentifier)
  * @see [tsIsIdentifier](http://vsc-base.org/#tsIsIdentifier)